backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.15.0
pydantic>=2.0.0
factorio-draftsman>=2.0.2
networkx>=3.0
django-cors-headers>=4.3.0
httpx>=0.27.0
